move redesign styles outside of components dir
